Move deprecated tests to DeprecatedDateFunctionsSuite
